Fix build errors by updating Supabase helpers
